Types Support
Implemented end-to-end support for types.  This support includes all
CRUD operations and an extra find method to enumerate all of the
types. This support has been tested at the unit level, as well as an
integration test for the web tier.  Currently there are migrations for
both H2 during testing and Postgres for production, however the
production migration has not been tested.