Points Support
Adds support for points within items.  As part of this, support was
also added for parsing point input as well as augmenting those points
with elevations. This support has been tested at the unit level, as
well as an integration test for the web tier.  Currently there are
migrations for both H2 during testing and Postgres for production,
however the production migration has not been tested.