Controller improvements
This change improves the controllers by specifying the character set of
the returned payloads (to support an accented e character) as well as
indicating that the void return type on DELETE is intentional.  Without
the indication, the delete method triggered a get of the type and then a
failure when the recently deleted type was no longer available.